requirements.txt
discord.py
python-dotenv
langchain
langchain_ollama
six